+++ packages/bytes/lib/index.js
@@ -73,7 +73,7 @@ function arrayify(value, options) {
         var hex = value.substring(2);
         if (hex.length % 2) {
             if (options.hexPad === "left") {
-                hex = "0x0" + hex.substring(2);
+                hex = "0" + hex;
             }
             else if (options.hexPad === "right") {
                 hex += "0";