+++ packages/providers/lib/json-rpc-provider.js
@@ -65,19 +65,47 @@ var _version_1 = require("./_version");
 var logger = new logger_1.Logger(_version_1.version);
 var base_provider_1 = require("./base-provider");
 var errorGas = ["call", "estimateGas"];
+function spelunk(value) {
+    if (value == null) {
+        return null;
+    }
+    // These *are* the droids we're looking for.
+    if (typeof (value.message) === "string" && value.message.match("reverted") && (0, bytes_1.isHexString)(value.data)) {
+        return { message: value.message, data: value.data };
+    }
+    // Spelunk further...
+    if (typeof (value) === "object") {
+        for (var key in value) {
+            var result = spelunk(value[key]);
+            if (result) {
+                return result;
+            }
+        }
+        return null;
+    }
+    // Might be a JSON string we can further descend...
+    if (typeof (value) === "string") {
+        try {
+            return spelunk(JSON.parse(value));
+        }
+        catch (error) { }
+    }
+    return null;
+}
 function checkError(method, error, params) {
     // Undo the "convenience" some nodes are attempting to prevent backwards
     // incompatibility; maybe for v6 consider forwarding reverts as errors
-    if (method === "call" && error.code === logger_1.Logger.errors.SERVER_ERROR) {
-        var e = error.error;
-        if (e && e.message.match("reverted") && (0, bytes_1.isHexString)(e.data)) {
-            return e.data;
+    if (method === "call") {
+        var result = spelunk(error);
+        if (result) {
+            return result.data;
         }
         logger.throwError("missing revert data in call exception", logger_1.Logger.errors.CALL_EXCEPTION, {
             error: error,
             data: "0x"
         });
     }
+    // @TODO: Should we spelunk for message too?
     var message = error.message;
     if (error.code === logger_1.Logger.errors.SERVER_ERROR && error.error && typeof (error.error.message) === "string") {
         message = error.error.message;